feat: support client and network comparisons
We now support doing a comparison of two or more client-only deployments.

To avoid duplication, a base class was created between a client and network deployment.

The database structure is so fundamentally different here that the previous migrations were just
removed in favour of starting again. Since it's just an internal tool we don't need to worry about
backwards compatibility.
