--- alembic/versions/055e50c9e047_add_smoke_test_results.py
@@ -1,37 +1,0 @@
-"""add smoke test results
-
-Revision ID: 055e50c9e047
-Revises: e7016646d944
-Create Date: 2024-12-30 17:15:55.222203
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '055e50c9e047'
-down_revision = 'e7016646d944'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('smoke_test_results',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('deployment_id', sa.Integer(), nullable=False),
-    sa.Column('results', sa.JSON(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['deployment_id'], ['deployments.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_smoke_test_results_id'), 'smoke_test_results', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_smoke_test_results_id'), table_name='smoke_test_results')
-    op.drop_table('smoke_test_results')
-    # ### end Alembic commands ###
