--- alembic/versions/473506d2ca0b_provide_client_smoke_test_results.py
@@ -1,37 +1,0 @@
-"""provide client smoke test results
-
-Revision ID: 473506d2ca0b
-Revises: bda96cb83924
-Create Date: 2025-04-10 21:22:05.867062
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '473506d2ca0b'
-down_revision = 'bda96cb83924'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('client_smoke_test_results',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('deployment_id', sa.Integer(), nullable=False),
-    sa.Column('results', sa.JSON(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['deployment_id'], ['client_deployments.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_client_smoke_test_results_id'), 'client_smoke_test_results', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_client_smoke_test_results_id'), table_name='client_smoke_test_results')
-    op.drop_table('client_smoke_test_results')
-    # ### end Alembic commands ###
