--- alembic/versions/95f7dbfb8326_region_column_for_deployment.py
@@ -1,28 +1,0 @@
-"""region column for deployment
-
-Revision ID: 95f7dbfb8326
-Revises: e6e242f022a1
-Create Date: 2025-04-04 17:03:47.322287
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '95f7dbfb8326'
-down_revision = 'e6e242f022a1'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('deployments', sa.Column('region', sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('deployments', 'region')
-    # ### end Alembic commands ###
