--- alembic/versions/bda96cb83924_provide_client_deployments.py
@@ -1,70 +1,0 @@
-"""provide client deployments
-
-Revision ID: bda96cb83924
-Revises: 95f7dbfb8326
-Create Date: 2025-04-10 19:59:49.467552
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'bda96cb83924'
-down_revision = '95f7dbfb8326'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('client_deployments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('workflow_run_id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('ant_version', sa.String(), nullable=True),
-    sa.Column('branch', sa.String(), nullable=True),
-    sa.Column('repo_owner', sa.String(), nullable=True),
-    sa.Column('chunk_size', sa.Integer(), nullable=True),
-    sa.Column('client_vm_count', sa.Integer(), nullable=False),
-    sa.Column('client_vm_size', sa.String(), nullable=False),
-    sa.Column('client_env', sa.String(), nullable=True),
-    sa.Column('evm_network_type', sa.String(), nullable=False),
-    sa.Column('evm_data_payments_address', sa.String(), nullable=True),
-    sa.Column('evm_payment_token_address', sa.String(), nullable=True),
-    sa.Column('evm_rpc_url', sa.String(), nullable=True),
-    sa.Column('network_id', sa.Integer(), nullable=True),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('related_pr', sa.Integer(), nullable=True),
-    sa.Column('triggered_at', sa.DateTime(), nullable=False),
-    sa.Column('run_id', sa.Integer(), nullable=False),
-    sa.Column('region', sa.String(), nullable=False),
-    sa.Column('provider', sa.String(), nullable=False),
-    sa.Column('wallet_secret_key', sa.String(), nullable=True),
-    sa.Column('environment_type', sa.String(), nullable=False),
-    sa.Column('disable_download_verifier', sa.Boolean(), nullable=True),
-    sa.Column('disable_performance_verifier', sa.Boolean(), nullable=True),
-    sa.Column('disable_random_verifier', sa.Boolean(), nullable=True),
-    sa.Column('disable_telegraf', sa.Boolean(), nullable=True),
-    sa.Column('disable_uploaders', sa.Boolean(), nullable=True),
-    sa.Column('expected_hash', sa.String(), nullable=True),
-    sa.Column('expected_size', sa.String(), nullable=True),
-    sa.Column('file_address', sa.String(), nullable=True),
-    sa.Column('initial_gas', sa.String(), nullable=True),
-    sa.Column('initial_tokens', sa.String(), nullable=True),
-    sa.Column('max_uploads', sa.Integer(), nullable=True),
-    sa.Column('network_contacts_url', sa.String(), nullable=True),
-    sa.Column('peer', sa.String(), nullable=True),
-    sa.Column('uploaders_count', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['workflow_run_id'], ['workflow_runs.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_client_deployments_id'), 'client_deployments', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_client_deployments_id'), table_name='client_deployments')
-    op.drop_table('client_deployments')
-    # ### end Alembic commands ###
