--- alembic/versions/ccd080ec96e5_new_private_node_columns.py
@@ -1,38 +1,0 @@
-"""new private node columns
-
-Revision ID: ccd080ec96e5
-Revises: e8d3bb9a87f8
-Create Date: 2025-02-13 19:12:49.506984
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'ccd080ec96e5'
-down_revision = 'e8d3bb9a87f8'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('deployments', sa.Column('full_cone_private_node_count', sa.Integer(), nullable=True))
-    op.add_column('deployments', sa.Column('full_cone_private_vm_count', sa.Integer(), nullable=True))
-    op.add_column('deployments', sa.Column('full_cone_nat_gateway_vm_size', sa.String(), nullable=True))
-    op.add_column('deployments', sa.Column('symmetric_private_node_count', sa.Integer(), nullable=True))
-    op.add_column('deployments', sa.Column('symmetric_private_vm_count', sa.Integer(), nullable=True))
-    op.add_column('deployments', sa.Column('symmetric_nat_gateway_vm_size', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('deployments', 'symmetric_nat_gateway_vm_size')
-    op.drop_column('deployments', 'symmetric_private_vm_count')
-    op.drop_column('deployments', 'symmetric_private_node_count')
-    op.drop_column('deployments', 'full_cone_nat_gateway_vm_size')
-    op.drop_column('deployments', 'full_cone_private_vm_count')
-    op.drop_column('deployments', 'full_cone_private_node_count')
-    # ### end Alembic commands ###
