--- alembic/versions/e6e242f022a1_record_the_client_and_node_environment_.py
@@ -1,30 +1,0 @@
-"""record the client and node environment variables for a deployment
-
-Revision ID: e6e242f022a1
-Revises: ccd080ec96e5
-Create Date: 2025-03-04 22:50:13.718181
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e6e242f022a1'
-down_revision = 'ccd080ec96e5'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('deployments', sa.Column('client_env', sa.String(), nullable=True))
-    op.add_column('deployments', sa.Column('node_env', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('deployments', 'node_env')
-    op.drop_column('deployments', 'client_env')
-    # ### end Alembic commands ###
