--- alembic/versions/e7016646d944_initial_migration.py
@@ -1,104 +1,0 @@
-"""initial migration
-
-Revision ID: e7016646d944
-Revises: 
-Create Date: 2024-12-30 15:54:55.193415
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e7016646d944'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('workflow_runs',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('workflow_name', sa.String(), nullable=False),
-    sa.Column('branch_name', sa.String(), nullable=False),
-    sa.Column('network_name', sa.String(), nullable=False),
-    sa.Column('triggered_at', sa.DateTime(), nullable=False),
-    sa.Column('inputs', sa.JSON(), nullable=False),
-    sa.Column('run_id', sa.Integer(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_workflow_runs_id'), 'workflow_runs', ['id'], unique=False)
-    op.create_table('deployments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('workflow_run_id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('ant_version', sa.String(), nullable=True),
-    sa.Column('antnode_version', sa.String(), nullable=True),
-    sa.Column('antctl_version', sa.String(), nullable=True),
-    sa.Column('branch', sa.String(), nullable=True),
-    sa.Column('repo_owner', sa.String(), nullable=True),
-    sa.Column('chunk_size', sa.Integer(), nullable=True),
-    sa.Column('antnode_features', sa.String(), nullable=True),
-    sa.Column('peer_cache_node_count', sa.Integer(), nullable=True),
-    sa.Column('generic_node_count', sa.Integer(), nullable=False),
-    sa.Column('private_node_count', sa.Integer(), nullable=True),
-    sa.Column('enable_downloaders', sa.Boolean(), nullable=True),
-    sa.Column('uploader_count', sa.Integer(), nullable=True),
-    sa.Column('peer_cache_vm_count', sa.Integer(), nullable=True),
-    sa.Column('generic_vm_count', sa.Integer(), nullable=False),
-    sa.Column('private_vm_count', sa.Integer(), nullable=True),
-    sa.Column('client_vm_count', sa.Integer(), nullable=True),
-    sa.Column('peer_cache_node_vm_size', sa.String(), nullable=True),
-    sa.Column('generic_node_vm_size', sa.String(), nullable=False),
-    sa.Column('private_node_vm_size', sa.String(), nullable=True),
-    sa.Column('client_vm_size', sa.String(), nullable=True),
-    sa.Column('evm_network_type', sa.String(), nullable=False),
-    sa.Column('rewards_address', sa.String(), nullable=False),
-    sa.Column('max_log_files', sa.Integer(), nullable=True),
-    sa.Column('max_archived_log_files', sa.Integer(), nullable=True),
-    sa.Column('evm_data_payments_address', sa.String(), nullable=True),
-    sa.Column('evm_payment_token_address', sa.String(), nullable=True),
-    sa.Column('evm_rpc_url', sa.String(), nullable=True),
-    sa.Column('related_pr', sa.Integer(), nullable=True),
-    sa.Column('network_id', sa.Integer(), nullable=True),
-    sa.Column('triggered_at', sa.DateTime(), nullable=False),
-    sa.Column('run_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['workflow_run_id'], ['workflow_runs.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_deployments_id'), 'deployments', ['id'], unique=False)
-    op.create_table('comparisons',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('ref_id', sa.Integer(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('thread_link', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('ref_label', sa.String(), nullable=True),
-    sa.Column('passed', sa.Boolean(), nullable=True),
-    sa.ForeignKeyConstraint(['ref_id'], ['deployments.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_comparisons_id'), 'comparisons', ['id'], unique=False)
-    op.create_table('comparison_deployments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('comparison_id', sa.Integer(), nullable=False),
-    sa.Column('deployment_id', sa.Integer(), nullable=False),
-    sa.Column('label', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['comparison_id'], ['comparisons.id'], ),
-    sa.ForeignKeyConstraint(['deployment_id'], ['deployments.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('comparison_deployments')
-    op.drop_index(op.f('ix_comparisons_id'), table_name='comparisons')
-    op.drop_table('comparisons')
-    op.drop_index(op.f('ix_deployments_id'), table_name='deployments')
-    op.drop_table('deployments')
-    op.drop_index(op.f('ix_workflow_runs_id'), table_name='workflow_runs')
-    op.drop_table('workflow_runs')
-    # ### end Alembic commands ###
