--- alembic/versions/e8d3bb9a87f8_add_a_description_to_deployments.py
@@ -1,28 +1,0 @@
-"""add a description to deployments
-
-Revision ID: e8d3bb9a87f8
-Revises: 055e50c9e047
-Create Date: 2025-01-24 09:24:49.634733
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e8d3bb9a87f8'
-down_revision = '055e50c9e047'
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('deployments', sa.Column('description', sa.String(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('deployments', 'description')
-    # ### end Alembic commands ###
