+++ runner/models.py
@@ -1,11 +1,12 @@
 from dataclasses import dataclass
 from datetime import datetime
 from typing import Optional
-from sqlalchemy import Column, Integer, String, DateTime, JSON, Boolean, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, JSON, Boolean, ForeignKey, Enum as SqlEnum
 from sqlalchemy.orm import relationship, mapped_column
 from sqlalchemy.orm.collections import attribute_mapped_collection
 from sqlalchemy.orm import backref
 from .database import Base
+from enum import Enum
 
 class WorkflowRun(Base):
     __tablename__ = "workflow_runs"
@@ -18,12 +19,30 @@ class WorkflowRun(Base):
     inputs = Column(JSON, nullable=False)
     run_id = Column(Integer, nullable=False)
 
-class Deployment(Base):
-    __tablename__ = "deployments"
+class DeploymentType(Enum):
+    NETWORK = "network"
+    CLIENT = "client"
+
+class BaseDeployment(Base):
+    __tablename__ = "base_deployments"
 
     id = Column(Integer, primary_key=True, index=True)
+    deployment_type = Column(SqlEnum(DeploymentType), nullable=False)
     workflow_run_id = Column(Integer, ForeignKey("workflow_runs.id"), nullable=False)
     name = Column(String, nullable=False)
+    triggered_at = Column(DateTime, nullable=False)
+    run_id = Column(Integer, nullable=False)
+    description = Column(String)
+    region = Column(String, nullable=False, default="lon1")
+    
+    __mapper_args__ = {
+        "polymorphic_on": deployment_type,
+    }
+
+class NetworkDeployment(BaseDeployment):
+    __tablename__ = "network_deployments"
+
+    id = Column(Integer, ForeignKey("base_deployments.id"), primary_key=True)
     ant_version = Column(String)
     antnode_version = Column(String)
     antctl_version = Column(String)
@@ -53,28 +72,23 @@ class Deployment(Base):
     evm_rpc_url = Column(String)
     related_pr = Column(Integer)
     network_id = Column(Integer)
-    description = Column(String)
-    triggered_at = Column(DateTime, nullable=False)
-    run_id = Column(Integer, nullable=False)
-    region = Column(String, nullable=False, default="lon1")
-
     client_env = Column(String)
     node_env = Column(String)
-    
     full_cone_private_node_count = Column(Integer)
     full_cone_private_vm_count = Column(Integer)
     full_cone_nat_gateway_vm_size = Column(String)
-
     symmetric_private_node_count = Column(Integer)
     symmetric_private_vm_count = Column(Integer)
     symmetric_nat_gateway_vm_size = Column(String)
+    
+    __mapper_args__ = {
+        "polymorphic_identity": DeploymentType.NETWORK,
+    }
 
-class ClientDeployment(Base):
+class ClientDeployment(BaseDeployment):
     __tablename__ = "client_deployments"
 
-    id = Column(Integer, primary_key=True, index=True)
-    workflow_run_id = Column(Integer, ForeignKey("workflow_runs.id"), nullable=False)
-    name = Column(String, nullable=False)
+    id = Column(Integer, ForeignKey("base_deployments.id"), primary_key=True)
     ant_version = Column(String)
     branch = Column(String)
     repo_owner = Column(String)
@@ -87,11 +101,7 @@ class ClientDeployment(Base):
     evm_payment_token_address = Column(String)
     evm_rpc_url = Column(String)
     network_id = Column(Integer)
-    description = Column(String)
     related_pr = Column(Integer)
-    triggered_at = Column(DateTime, nullable=False)
-    run_id = Column(Integer, nullable=False)
-    region = Column(String, nullable=False, default="lon1")
     provider = Column(String, nullable=False, default="digital-ocean")
     wallet_secret_key = Column(String)
     environment_type = Column(String, nullable=False)
@@ -109,34 +119,37 @@ class ClientDeployment(Base):
     network_contacts_url = Column(String)
     peer = Column(String)
     uploaders_count = Column(Integer, nullable=False, default=1)
+    
+    __mapper_args__ = {
+        "polymorphic_identity": DeploymentType.CLIENT,
+    }
 
 class ComparisonDeployment(Base):
-    """Association table for many-to-many relationship between comparisons and test deployments"""
     __tablename__ = "comparison_deployments"
 
     id = Column(Integer, primary_key=True)
     comparison_id = Column(Integer, ForeignKey("comparisons.id"), nullable=False)
-    deployment_id = Column(Integer, ForeignKey("deployments.id"), nullable=False)
+    deployment_id = Column(Integer, ForeignKey("base_deployments.id"), nullable=False)
     label = Column(String)
     
     comparison = relationship("Comparison", back_populates="test_deployments")
-    deployment = relationship("Deployment")
+    deployment = relationship("BaseDeployment")
 
 class Comparison(Base):
     __tablename__ = "comparisons"
 
     id = Column(Integer, primary_key=True, index=True)
-    ref_id = Column(Integer, ForeignKey("deployments.id"), nullable=False)
+    ref_id = Column(Integer, ForeignKey("base_deployments.id"), nullable=False)
+    deployment_type = Column(SqlEnum(DeploymentType), nullable=False)
     description = Column(String)
     thread_link = Column(String)
     created_at = Column(DateTime, nullable=False)
     ref_label = Column(String)
     passed = Column(Boolean)
 
-    ref_deployment = relationship("Deployment", foreign_keys=[ref_id])
+    ref_deployment = relationship("BaseDeployment", foreign_keys=[ref_id])
     test_deployments = relationship("ComparisonDeployment", back_populates="comparison")
     
-    # Instead of association_proxy, we can access deployments through the relationship
     @property
     def test_environments(self):
         return [(cd.deployment, cd.label) for cd in self.test_deployments]
@@ -145,11 +158,11 @@ class SmokeTestResult(Base):
     __tablename__ = "smoke_test_results"
 
     id = Column(Integer, primary_key=True, index=True)
-    deployment_id = Column(Integer, ForeignKey("deployments.id"), nullable=False)
+    deployment_id = Column(Integer, ForeignKey("base_deployments.id"), nullable=False)
     results = Column(JSON, nullable=False)
     created_at = Column(DateTime, nullable=False)
 
-    deployment = relationship("Deployment", backref="smoke_test_results")
+    deployment = relationship("BaseDeployment", backref="smoke_test_results")
 
 class ClientSmokeTestResult(Base):
     __tablename__ = "client_smoke_test_results"
